Avoid generating invalid USE flag combinations
Consider ignored (through `ignoreprefix`) USE flags that are enabled on
the system as "always on" in `REQUIRED_USE` checks of possible
combinations. This makes us avoid generating invalid combinations.

For example, on a system with `FEATURES=test`, a package with `test? (
foo bar )` will always generate a USE line with `foo bar` in it.

Also, fix the call to portage's `check_required_use()`: it expects a
list, not a string.

fixes #52